ci: add check for MethodByName
It's easy to introduce a dependency that uses MethodByName that forces
the compiler into conservative mode. Write a check based on go tool nm
to verify that basic import of the package doesn't create a binary size
problem.

Also use `go build` instead of `go vet` for cross-compatibility, because
we can allow tests to use newer Go features.

Change-Id: I00b488d399e1c16cfd4e501c18eb5122ee24cd74